+++ core/src/test/java/io/fabric8/maven/core/service/GroupArtifactVersionResolverServiceMavenImplTest.java
@@ -20,7 +20,6 @@ import java.util.Collections;
 
 import mockit.Expectations;
 import mockit.Mocked;
-import mockit.integration.junit4.JMockit;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.resolver.ArtifactResolutionRequest;
 import org.apache.maven.artifact.resolver.ArtifactResolutionResult;
@@ -28,12 +27,10 @@ import org.apache.maven.project.MavenProject;
 import org.apache.maven.repository.RepositorySystem;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-@RunWith(JMockit.class)
 public class GroupArtifactVersionResolverServiceMavenImplTest {
 
     @Mocked