+++ enricher/fabric8/src/test/java/io/fabric8/maven/enricher/fabric8/PrometheusEnricherTest.java
@@ -29,14 +29,11 @@ import io.fabric8.maven.enricher.api.MavenEnricherContext;
 import io.fabric8.maven.enricher.api.Kind;
 import mockit.Expectations;
 import mockit.Mocked;
-import mockit.integration.junit4.JMockit;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 
-@RunWith(JMockit.class)
 public class PrometheusEnricherTest {
 
     @Mocked