+++ enricher/fabric8/src/test/java/io/fabric8/maven/enricher/fabric8/SpringBoot2HealthCheckEnricherTest.java
@@ -15,10 +15,7 @@
  */
 package io.fabric8.maven.enricher.fabric8;
 
-import mockit.integration.junit4.JMockit;
-import org.junit.runner.RunWith;
 
-@RunWith(JMockit.class)
 public class SpringBoot2HealthCheckEnricherTest extends AbstractSpringBootHealthCheckEnricherTestSupport {
 
     @Override