+++ enricher/osio/src/test/java/io/fabric8/maven/enricher/osio/SpaceLabelEnricherTest.java
@@ -28,15 +28,12 @@ import mockit.Expectations;
 import mockit.Mock;
 import mockit.MockUp;
 import mockit.Mocked;
-import mockit.integration.junit4.JMockit;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 
-@RunWith(JMockit.class)
 public class SpaceLabelEnricherTest {
 
     @Mocked