+++ enricher/standard/src/test/java/io/fabric8/maven/enricher/standard/ConfigMapEnricherTest.java
@@ -26,13 +26,10 @@ import java.util.HashMap;
 import java.util.Map;
 import mockit.Expectations;
 import mockit.Mocked;
-import mockit.integration.junit4.JMockit;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@RunWith(JMockit.class)
 public class ConfigMapEnricherTest {
 
     @Mocked