+++ enricher/standard/src/test/java/io/fabric8/maven/enricher/standard/DefaultControllerEnricherTest.java
@@ -30,11 +30,9 @@ import java.util.Collections;
 import java.util.TreeMap;
 import mockit.Expectations;
 import mockit.Mocked;
-import mockit.integration.junit4.JMockit;
 import org.apache.maven.project.MavenProject;
 import org.hamcrest.Matchers;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
@@ -43,7 +41,6 @@ import static org.junit.Assert.assertThat;
  * @author kamesh
  * @since 08/05/17
  */
-@RunWith(JMockit.class)
 public class DefaultControllerEnricherTest {
 
     @Mocked