+++ enricher/standard/src/test/java/io/fabric8/maven/enricher/standard/VolumePermissionEnricherTest.java
@@ -36,15 +36,12 @@ import io.fabric8.maven.core.model.Configuration;
 import io.fabric8.maven.enricher.api.MavenEnricherContext;
 import mockit.Expectations;
 import mockit.Mocked;
-import mockit.integration.junit4.JMockit;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-@RunWith(JMockit.class)
 public class VolumePermissionEnricherTest {
 
     @Mocked