+++ plugin/src/test/java/io/fabric8/maven/plugin/enricher/EnricherManagerTest.java
@@ -32,9 +32,7 @@ import java.util.Optional;
 
 import mockit.Expectations;
 import mockit.Mocked;
-import mockit.integration.junit4.JMockit;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -44,7 +42,6 @@ import static org.junit.Assert.assertTrue;
  * @author roland
  * @since 23/09/16
  */
-@RunWith(JMockit.class)
 public class EnricherManagerTest {
 
     @Mocked