+++ CHANGELOG.md
@@ -10,6 +10,13 @@ We use semantic versioning in some slight variation until our feature set has st
 
 After this we will switch probably to real [Semantic Versioning 2.0.0](http://semver.org/)
 
+### 4.0.0-M2-SNAPSHOT
+* Fix 1326: Fixes overridding of Selector Label by the project enrichers entries.
+* Fix 839: Sets Spring Boot generator color config property as String (http://docs.spring.io/spring-boot/docs/current/api/org/springframework/boot/ansi/AnsiOutput.Enabled.html)
+* Fix 1412: mvn deploy fails when using a Dockerfile during S2I build
+* Fix 796: Remove workaround to produce both .yaml and .json files
+* Fix 1425: Added metadata visitors for imagestreams, build and buildconfig.
+
 ### 3.5-SNAPSHOT
 * Fix 1021: Avoids empty deployment selector value in generated yaml resource
 * Fix 1383: Check if instanceof Openshift before casting and handle failures.
@@ -38,6 +45,7 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 * Fix 977: Allows user to define ConfigMap from XML and annotation
 * Fix 314: Do not execute GitEnricher if not Git repo
 * Fix 714: Merges the functionality of IANAServicePortNameEnricher and PortNameEnricher.
+* Fix 963: add support for Pod resource type in yaml config.
 
 ### 3.5.41 (2018-08-01)
 * Feature 1032: Improvements of the Vert.x Generator and enrichers